requirements.txt
markdownify
smolagents==1.13.0
requests
duckduckgo_search
pandas
pypdf
gradio
pyyaml
litellm
pinecone
google-generativeai
PyPDF2
python-dotenv
langchain-text-splitters
